ai: Move count tracer to a new file
