+++ libs/ai/include/ai.h
@@ -1,4 +1,3 @@
 #pragma once
 #include "ai_blackboard.h"
 #include "ai_eval.h"
-#include "ai_tracer.h"
