+++ libs/ai/include/ai_tracer.h
@@ -1,6 +1,5 @@
 #pragma once
 #include "ai_result.h"
-#include "core_types.h"
 
 // Forward declare from 'asset_behavior.h'.
 typedef struct sAssetBehavior AssetBehavior;
@@ -18,15 +17,3 @@ typedef struct sAiTracer {
  * Destroy a tracer instance.
  */
 void ai_tracer_destroy(AiTracer*);
-
-/**
- * Simple tracer that counts the amount of node evaluations.
- * NOTE: Returned by value, doesn't need to be destroyed.
- */
-
-typedef struct {
-  AiTracer api;
-  u32      count;
-} AiTracerCount;
-
-AiTracerCount ai_tracer_count();
