+++ libs/ai/include/ai_tracer_count.h
@@ -1,0 +1,15 @@
+#pragma once
+#include "ai_tracer.h"
+#include "core_types.h"
+
+/**
+ * Simple tracer that counts the amount of node evaluations.
+ * NOTE: Returned by value, doesn't need to be destroyed.
+ */
+
+typedef struct {
+  AiTracer api;
+  u32      count;
+} AiTracerCount;
+
+AiTracerCount ai_tracer_count();
