+++ libs/ai/test/test_node_selector.c
@@ -1,4 +1,5 @@
 #include "ai.h"
+#include "ai_tracer_count.h"
 #include "asset_behavior.h"
 #include "check_spec.h"
 #include "core_alloc.h"
